refactor : 정치인 상세정보 페이지 컴포넌트 분리 및 리팩토링
