+++ src/pages/politician/brief_info_container/BriefInfoContainer.module.scss
@@ -1,0 +1,76 @@
+.profileBox {
+  width: 100%;
+  height: 30%;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  gap: 10px;
+
+  &__profile {
+    height: 100%;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: start;
+    align-items: center;
+    gap: 5px;
+
+    background-color: white;
+    border-radius: 10px;
+    padding-left: 7px;
+    padding-top: 7px;
+    padding-bottom: 7px;
+    padding-right: 2px;
+    @include mobile {
+      width: 50%;
+    }
+    @include non-mobile {
+      flex-basis: 0;
+      flex-grow: 2;
+    }
+    @include desktop {
+      flex-basis: 0;
+      flex-grow: 6;
+    }
+  }
+
+  &__imageSection {
+    position: relative;
+    height: 100%;
+
+    border-radius: 10px;
+    background-color: white;
+
+    display: flex;
+    justify-content: start;
+    align-items: start;
+
+    overflow: hidden;
+    @include mobile {
+      width: 50%;
+    }
+    @include non-mobile {
+      flex-basis: 0;
+      flex-grow: 1;
+    }
+    @include desktop {
+      flex-basis: 0;
+      flex-grow: 1;
+    }
+
+    &__image {
+      width: 100%;
+      height: 100%;
+      border-radius: 10px;
+      object-fit: contain;
+    }
+    &__icon {
+      position: absolute;
+      left: 3px;
+
+      cursor: pointer;
+    }
+  }
+}
