+++ src/pages/politician/brief_info_container/BriefInfoItem.module.scss
@@ -1,0 +1,39 @@
+.dataBox {
+  width: 100%;
+  height: 20%;
+
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+
+  &__label {
+    width: 30%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    background-color: $color-purple-500;
+    color: white;
+    border-radius: 10px;
+    z-index: 2;
+
+    font-size: clamp(11px, 1vw, 22px);
+  }
+  &__data {
+    width: 70%;
+    height: 90%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin-left: -5px;
+    z-index: 1;
+    padding: 7px;
+
+    background-color: $color-purple-300;
+    border-top-right-radius: 10px;
+    border-bottom-right-radius: 10px;
+
+    font-size: clamp(11px, 1vw, 22px);
+  }
+}
