+++ src/pages/politician/styles/politician.module.scss
@@ -17,184 +17,7 @@
     gap: 10px;
 
     background-color: $color-gray-200;
-
-    &__profileBox {
-      width: 100%;
-      height: 30%;
-
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      align-items: center;
-      gap: 10px;
-
-      &__profile {
-        height: 100%;
-
-        display: flex;
-        flex-direction: column;
-        justify-content: start;
-        align-items: center;
-        gap: 5px;
-
-        background-color: white;
-        border-radius: 10px;
-        padding-left: 7px;
-        padding-top: 7px;
-        padding-bottom: 7px;
-        padding-right: 2px;
-        @include mobile {
-          width: 50%;
-        }
-        @include non-mobile {
-          flex-basis: 0;
-          flex-grow: 2;
-        }
-        @include desktop {
-          flex-basis: 0;
-          flex-grow: 6;
-        }
-
-        &__dataBox {
-          width: 100%;
-          height: 20%;
-
-          display: flex;
-          flex-direction: row;
-          align-items: center;
-
-          &__label {
-            width: 30%;
-            height: 100%;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-
-            background-color: $color-purple-500;
-            color: white;
-            border-radius: 10px;
-            z-index: 2;
-
-            font-size: clamp(11px, 1vw, 22px);
-          }
-          &__data {
-            width: 70%;
-            height: 90%;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            margin-left: -5px;
-            z-index: 1;
-            padding: 7px;
-
-            background-color: $color-purple-300;
-            border-top-right-radius: 10px;
-            border-bottom-right-radius: 10px;
-
-            font-size: clamp(11px, 1vw, 22px);
-          }
-        }
-      }
-
-      &__imageSection {
-        position: relative;
-        height: 100%;
-
-        border-radius: 10px;
-        background-color: white;
-
-        display: flex;
-        justify-content: start;
-        align-items: start;
-
-        overflow: hidden;
-        @include mobile {
-          width: 50%;
-        }
-        @include non-mobile {
-          flex-basis: 0;
-          flex-grow: 1;
-        }
-        @include desktop {
-          flex-basis: 0;
-          flex-grow: 1;
-        }
-
-        &__image {
-          width: 100%;
-          height: 100%;
-          border-radius: 10px;
-          object-fit: contain;
-        }
-        &__icon {
-          position: absolute;
-          left: 3px;
-
-          cursor: pointer;
-        }
-      }
-    }
-
-    &__infoBox {
-      width: 100%;
-      height: 70%;
-
-      display: flex;
-      flex-direction: column;
-
-      background-color: white;
-      border-radius: 10px;
-      padding: 7px;
-      gap: 7px;
-
-      &__categoryBox {
-        overflow: hidden;
-        user-select: none;
-        &__window {
-          // width: 100%;
-          width: calc(20% * 6);
-          height: 2rem;
-
-          display: flex;
-          flex-direction: row;
-          justify-content: start;
-          gap: 10px;
-
-          &__category {
-            height: 2rem;
-
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            flex: 0 0 19%;
-
-            background-color: $color-purple-300;
-            border-radius: 10px;
-            font-size: 0.9rem;
-
-            cursor: pointer;
-          }
-        }
-      }
-
-      &__view {
-        width: 100%;
-        flex-grow: 1;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-
-        background-color: $color-gray-200;
-        border-radius: 10px;
-      }
-    }
   }
-}
-
-.active {
-  background-color: $color-purple-500;
-  color: white;
 }
 
 .nonObserve {
